refactor : index.tsx 코드 정리
